"Normalize" our combined values so they don't go off the graph
Why normalize to a max of 2? I have no idea... that's just what worked.

Best guess is that because in the `toPoints` function I'm adjusting the values sort of randomly to make them also fit on to the graph, that I have to have another, matching, magic number when adding them all together.

SO, basically because i've adjusted with magic numbers there, it's no surprise that I have to adjust by more magic numbers when combining the waves.